feat: Добавлен контроллер админа с get запросом на получение списка всех пользователей, put запросом на одобрение аккаунта и post запросом на списание долга

+++ BLL/Models/TransactionModel.cs
@@ -19,7 +19,7 @@ namespace BLL.Models
         public TransactionModel(Transaction r) { 
             Id = r.Id;
             User = r.User;
-            Admin = r.Admin;
+            Admin = (int)r.Admin;
             Price = r.Price;
             Date = r.Date;
         }
